% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-loci.R
\name{read_loci}
\alias{read_loci}
\title{Read MSI locus tables from CLC reports in tidy format}
\usage{
read_loci(filepath)
}
\arguments{
\item{filepath}{The filepath to the CLC MSI report Excel}
}
\value{
A dataframe of locus information in tidy format.
}
\description{
\code{read_loci} reads in the locus genotype tables from a CLC MSI report and
converts them into a tidy format for easier data analysis.
}
\section{Initial data format}{

MSI reports exported from CLC contain information on the genotype of
each tested locus, however the data structure is difficult to work with.
\itemize{
\item The name of the locus marker is on one row
\item The stability of the marker ("Stable" or "Unstable") is on the next row
\item Then there is a genotype table including the repeat lengths of the locus and
the percentage of reads in the sample and baseline associated with each length.
\item There is a gap of two rows before the same information for the next
marker, which may have different repeat lengths associated with it.
}
}

\section{What \code{read_loci} does}{


\code{read_loci} rearranges the initial format using 3 helper functions.
\itemize{
\item \code{read_msi_locus_section}: first, the section of the Excel sheet
containing the locus tables is read in as a single table.
\item \code{mutate_msi_row_numbers}: the numbers of each row are annotated.
\item \code{join_msi_locus_tables}: the three pieces of information for each marker (the name,
stability and repeat table) are joined together based on their row numbers
with : the marker name will be 2 rows above the repeat
table, and the marker stability will be 1 row above.
}
}

\section{Output data format}{


This produces a dataframe which includes the information in the repeat marker
table, with the marker name and stability as additional columns.
}

\section{Usage}{


\code{read-loci} can be combined with \code{purr::map} and \code{purrr::list_rbind} to
combine marker tables from multiple CLC MSI reports.
}

