\name{tidymsi-package}
\alias{tidymsi-package}
\alias{tidymsi}
\docType{package}
\title{
\packageTitle{tidymsi}
}
\description{
\packageDescription{tidymsi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tidymsi}
\packageIndices{tidymsi}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tidymsi}

Maintainer: \packageMaintainer{tidymsi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
